meow

+++ include/miku/devices/distance.h
@@ -10,18 +10,18 @@ enum class Orientation {
 
 namespace miku {
 class Distance : public pros::Distance {
-    double data = -1;
+    float data = -1;
     bool valid = false;
     bool enabled = true;
 public:
-    double offset_x, offset_y;
+    float offset_x, offset_y;
     Orientation orientation;
-    Distance(const std::uint8_t port, double offset_x, double offset_y, Orientation orientation) 
+    Distance(const std::uint8_t port, float offset_x, float offset_y, Orientation orientation) 
         : pros::Distance(port), offset_x(offset_x), offset_y(offset_y), orientation(orientation) {}
     void set_enabled(bool enable) { enabled = enable; }
     void update_reading();
     bool get_enabled() { return enabled; }
     bool get_valid() { return valid; }
-    double get_reading() { return data; }
+    float get_reading() { return data; }
 };
 }
