+++ include/miku/lut.h
@@ -4,12 +4,12 @@
 
 class LookupTable {
     private:
-        std::vector<std::pair<double, double>> table;
+        std::vector<std::pair<float, float>> table;
     public:
         LookupTable() {};
-        LookupTable(std::vector<std::pair<double, double>> table) : table(table) {};
-        LookupTable(std::initializer_list<std::pair<double, double>> table) : table(table) {};
-        double get_value(double key);
+        LookupTable(std::vector<std::pair<float, float>> table) : table(table) {};
+        LookupTable(std::initializer_list<std::pair<float, float>> table) : table(table) {};
+        float get_value(float key);
 };
 
 void tune_lut_drive();
