+++ include/miku/pid.h
@@ -4,35 +4,35 @@
 #include "miku/lut.h"
 
 struct PIDGains {
-    double kP;
-    double kI;
-    double kD;
+    float kP;
+    float kI;
+    float kD;
 
-    PIDGains(double kP, double kI, double kD) : kP(kP), kI(kI), kD(kD) {}
+    PIDGains(float kP, float kI, float kD) : kP(kP), kI(kI), kD(kD) {}
     PIDGains() : kP(0.0), kI(0.0), kD(0.0) {}
 };
 
 struct FeedforwardGains {
-    double kS;
-    double kV;
-    double kA;
+    float kS;
+    float kV;
+    float kA;
 
-    FeedforwardGains(double kS, double kV, double kA) : kS(kS), kV(kV), kA(kA) {}
+    FeedforwardGains(float kS, float kV, float kA) : kS(kS), kV(kV), kA(kA) {}
     FeedforwardGains() : kS(0.0), kV(0.0), kA(0.0) {}
 };
 
 /**
  * @param Gains gains The PID gains
- * @param double windup_range The range at which to limit the integral term (default 0, no limit)
+ * @param float windup_range The range at which to limit the integral term (default 0, no limit)
  * @param bool sign_flip_reset Whether to reset the integral term when the error changes sign (default true)
  * @param bool trapezoidal Whether to use trapezoidal integration for the integral term (default true)
 */
 class PID {
     public:
-        PID(PIDGains pid_gains, double windup_range = 0, bool sign_flip_reset = true, bool trapezoidal = true);
+        PID(PIDGains pid_gains, float windup_range = 0, bool sign_flip_reset = true, bool trapezoidal = true);
         PID();
 
-        double update(double error);
+        float update(float error);
 
         void reset();
 
@@ -42,10 +42,10 @@ class PID {
         PIDGains pid_gains;
 
         // optimizations
-        const double windup_range;
+        const float windup_range;
         const bool sign_flip_reset;
         const bool trapezoidal;
 
-        double integral = 0;
-        double prevError = 0;
+        float integral = 0;
+        float prevError = 0;
 };
